requirements.txt
langchain>=0.1.0
chromadb>=0.4.0
pytesseract>=0.3.10
Pillow>=10.0.0
pytest>=7.0.0
openai>=1.0.0
streamlit>=1.30.0
requests>=2.31.0
